Allears STIM LIB Application TEST
Allears STIM LIB Application TEST

Allears STIM LIB Application TEST

Mode change

Update stim_lib_stim_cfg_st.c

save

Update td_stim_mask_param_table.c

Update main.c

save

Update stim_lib.c

Update stim_lib.c

save

save

Update stim_lib.c

FUCK

save

Update stim_lib.c

save

save

save

Update main.c

FUCK

fuck

save

fuck

fuck

FUCK

fuck

FUCK

save

save

Update stim_lib.c

save

Update .cproject

Update stim_lib_volt.c

+++ 48 PIN/allears_EVK_BULK_48PIN/Core/Src/user_app/general/td_schedule.c
@@ -75,9 +75,9 @@ void td_Schedule(void)
  * */
 void td_Group_Pulse_Mode_Control_Scheduler(void)
 {
-	/* TIME VALUE = 100ms * 10 */
-	uint16_t pwm_disable_tim = (TD_RAW_GROUP_PULSE_DISABLE_TIME * 10) - 1;
-	uint16_t pwm_enable_tim = (TD_RAW_GROUP_PULSE_ENABLE_TIME * 10);
+	/* TIME VALUE = 100ms * 10 >> 1ms * 1000 */
+	uint32_t pwm_disable_tim = (TD_RAW_GROUP_PULSE_DISABLE_TIME * 10) - 1;
+	uint32_t pwm_enable_tim = (TD_RAW_GROUP_PULSE_ENABLE_TIME * 10);
 
 #if 1
 	/* CHANGE Group Pulse MODE */
@@ -94,6 +94,7 @@ void td_Group_Pulse_Mode_Control_Scheduler(void)
 		/* IF :: STATUS SLEEP >> AWAKE */
 		if (td_Get_Current_GP_State() == td_gp_sleep && gp_ctrl_cnt_tim16 == pwm_disable_tim)
 		{
+			//TIM2->CCER = 0x1112;
 			stimLib_stimStart();
 			gp_stat = td_gp_awake;
 			gp_ctrl_cnt_tim16 = 0;
@@ -102,6 +103,7 @@ void td_Group_Pulse_Mode_Control_Scheduler(void)
 		/* IF :: STATUS AWAKE >> SLEEP */
 		else if (td_Get_Current_GP_State() == td_gp_awake && gp_ctrl_cnt_tim16 == pwm_enable_tim)
 		{
+			//TIM2->CCER = 0x0000;
 			stimLib_stimPause();
 			gp_stat = td_gp_sleep;
 			gp_ctrl_cnt_tim16 = 0;
